requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==2.5.1
python-dotenv==1.0.0
pymysql==1.0.3